an ok auth flow
